move TblEntry to extra class before redesign
