+++ src/dat/Tbl.h
@@ -9,49 +9,14 @@
 
 // Local
 #include "kaitai/file_tbl.h"
+#include "TblEntry.h"
 
 // System
 #include <memory>
 
 namespace dat {
 
-class TblEntry
-{
-public:
-  TblEntry()
-  {
-    shortcut_pos = -1;
-    shortcut = ' ';
-  }
 
-  // TODO: this was a bad decision: make this functional access to add some logging functions...
-  std::string name1;
-  std::string name2;
-  std::string name3;
-  int shortcut_pos;
-  std::string shortcut;
-
-
-  void removeSpaces()
-  {
-    size_t pos;
-    while ((pos = name1.find("  ")) != std::string::npos)
-    {
-      name1 = name1.replace(pos, 2, " ");
-    }
-
-    while ((pos = name2.find("  ")) != std::string::npos)
-    {
-      name2 = name2.replace(pos, 2, " ");
-    }
-
-    while ((pos = name3.find("  ")) != std::string::npos)
-    {
-      name3 = name3.replace(pos, 2, " ");
-    }
-  }
-
-};
 
 class Tbl
 {
